# docs/api-server/openapi.yaml
openapi: "3.1.0"
info:
  title: AgentClash API
  version: 1.0.0
  description: >
    B2B AI agent evaluation platform. Pits agent deployments against challenge
    packs, with replay inspection, scoring, comparison, and release gating.
  license:
    name: FSL-1.1-MIT
    url: https://github.com/agentclash/agentclash/blob/main/LICENSE
  contact:
    url: https://github.com/agentclash/agentclash
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/workspaces/{workspaceID}/baselines/{name}:
    get:
      operationId: getBaseline
      tags: [Baselines]
      summary: Get a baseline by name
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: name
          in: path
          required: true
          schema: { type: string }
          description: Baseline name (e.g. `main`).
      responses:
        "200":
          description: Baseline record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Baseline"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      operationId: upsertBaseline
      tags: [Baselines]
      summary: Create or overwrite a named baseline
      description: >
        Idempotent by `(workspace_id, name)` — a POST with an existing name
        replaces `pack_version_id`, `run_id`, and `scorecard_snapshot`.
        Baselines are moving bookmarks by design; users retrain "main" when
        their challenge pack evolves.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBaselineRequest"
      responses:
        "200":
          description: Baseline created or replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Baseline"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

# =============================================================================
# Components
# =============================================================================
components:
...
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Bearer token for authenticated API requests. Web sessions use a WorkOS JWT
        and CLI sessions use an AgentClash token prefixed with `clitok_`.
        In development mode, use custom headers instead:
        X-Agentclash-User-Id, X-Agentclash-WorkOS-User-Id,
        X-Agentclash-User-Email, X-Agentclash-User-Display-Name,
        X-Agentclash-Workspace-Memberships.
    hostedRunToken:
      type: http
      scheme: bearer
      description: >
        HMAC-SHA256 callback token issued when a hosted run is started.
        Not a JWT — the token encodes run_id and run_agent_id claims.
    artifactSignature:
      type: apiKey
      in: query
      name: signature
      description: >
        HMAC-SHA256 signature for artifact content URLs. Must be paired with
        an "expires" query parameter (unix timestamp).

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    WorkspaceID:
      name: workspaceID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    Baseline:
      type: object
      required: [id, workspace_id, name, pack_version_id, run_id, scorecard_snapshot, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        pack_version_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        scorecard_snapshot:
          type: object
          additionalProperties: true
          description: >
            Point-in-time scorecard captured by the client when the baseline
            was saved. Opaque JSON — clients treat it as the shape that
            `agentclash eval` emits under `.ranking`, but the backend does
            not parse it.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpsertBaselineRequest:
      type: object
      required: [pack_version_id, run_id]
      properties:
        pack_version_id:
          type: string
          format: uuid
          description: Published pack version the baseline was captured on.
        run_id:
          type: string
          format: uuid
          description: Run whose scorecard is being bookmarked.
        scorecard_snapshot:
          type: object
          additionalProperties: true
          description: Opaque client-supplied JSON to store alongside the baseline.